Add authentication to HomeController
Closes #3

In this commit, we require users to be authenticated before rendering the
`Home#index` page. If the user is not signed in, we redirect them to the login
page.

Other changes:
- in `spec/support/devise.rb`, include Devise helpers in certain specs so that
  we can use helper methods, e.g., `sign_in`
- in `config/environments/test.rb`, set the application to eager load. Without
  this, we were seeing the below Runtime error when running a spec with a
`user` created using FactoryBot

```ruby
Could not find a valid mapping for #<User id: 21, email:
"testuser1@email.com", created_at: "2025-08-21 09:25:00.972588000 +0000",
updated_at: "2025-08-21 09:25:00.972588000 +0000">
 ```