Move certain gems to the :development, :test group
In an earlier commit, I moved certain gems that I thought would only be used in
tests into a `:test` group.

However, this was leading the rails model generator to blow up, because when
generating the spec for the model we're creating, we load the RSpec
environment. This loads `spec_helper`, which requires all files in the
`spec/support` directory.

In order to prevent this blow up, we move gems that are referenced in
`spec/support` into the `:development, :test` group.