Allow users to create Group through UI (#12)
In this commit we add a Group form that allows users to create a `Group`
through the UI.

As part of this work, we require a user to be logged in before accessing any
group related routes.

Additionally, we add a presence validation for `:name` on the `Group` model and
re-render the form with an error if a user tries to create a `Group` without a
name.