Add users to the Group upon creation (#12)
In this commit, we allow the `current_user` to add one or more additional users
to the `Group` upon creation.

The list of available users currently includes all users, except the
`current_user`. The `current_user` automatically gets added to the `Group` on
save.

When the `Group` is created, redirect to the `Group` show page, which lists the
`Group`'s users.