Add conditional Balances section to Group show page (#15)
In this commit, we display a conditional Balances section on the Group show
page to show which group members owe or are owed money. This Balances section
as well as the Expenses section are conditionally rendered depending on whether
any expenses have been added to the group.

The balances are calculated by a `Balances` query object, which assumes an even
split of all expeses among group members. Each member's balance is the
difference of:
- each group member's share of expenses (Total group expenses / Number of group
  members)
- what each group member paid (Sum of group expenses belonging to the user)