Add form to create Expense (#16)
In this commit, we add a form for a user to create an expense associated with a
particular group and user.

We only allow a user to add expenses to a `Group` to which they belong and we
only allow users to select an owner of the `Expense` from among `User`s who
belong to their groups.

Currently, the `current_user` could possible select a `Group` from the Group
select and a User from the User select that does not belong to the selected
Group. This would lead to the creation of an invalid `Expense`, since the
`User` to which the `Expense` belongs would not be a member of the `Group` to
which the `Expense` belongs.

To get around this, we add a custom validation to `Expense` that requires that
the `Expense`'s `User` must also belong to the `Expense`'s `Group`.

In future work that implements React, we will refetch options for the "User"
select when the "Group" changes, so that users can only create a valid
`Expense` through the UI.  #Please enter the commit message for your changes.
Lines starting