Include users in groups query in HomeController#index
In the `HomeController#index` action, we have a query to get all the `groups`
for the `current_user`, which is used to display the logged-in user's groups on
the home page. The `GroupSerializer` also serializes each `Group`'s `Users`, in
order to display which users are in each `Group` on a group card.

As the query is curently written, this leads to an N+1 query problem, since a
separate query for each group's users gets made. In order to mitigate this, we
add `includes(:users)` in the query in the controller in order to load the
users for each group with one additional query.