Mount GroupsContainer after Turbo navigation event
I noticed that when clicking the "BillSp/it" logo in the navbar, the
`GroupCard`s were not showing up on the home page; however, visiting the `/`
path and triggering a full page load would cause the cards to mount.

This unwanted behavior is due to **Turbo Drive**, enabled by default in Rails
7+. It intercepts link clicks and loads pages via AJAX, replacing on the
`<body>` content. This causes issues with React components that are mounted on
page load, because the TS initialization code might not run again after a Turbo
navigation.

In short, clicking the link in the navbar was causing only the HTML on the page
to be replace; the React mounting code wasn't re-run, resulting in an empty
page.

One solution could be to diable Turbo on a per-link basis:

```ruby
<%= link_to(
  "BillSp/it",
  root_path,
  class: "text-indigo-300 text-3xl font-bold tracking-[0.1rem]",
  data: { turbo: false }
) %>
```

A more robust solution is to ensure that the React app gets mounted on the
Turbo navigation event.

In the `GroupsContainer.ts` entrypoint, we wrap the `mountComponent` function
in a `turbo:load` event listener so that the component gets re-mounted after
Turbo navigation.