Fix React not mounting after Turbo load problem
I'm continuing to see the problem where the React component doesn't mount after
the Turbo load event.

In this commit, we attempt to solve the problem by adding turbo:load event
listener in `app/javascript/entrypoints/application.ts` which searches the
document for the tag where the React component is supposed to be mounted, then
if it's found, mounts the component by calling the mounting function found in
that component's entrypoint.