Write new expense form in React (#30)
In a previous issue, we created a modal that appears on the group show page to
make a new expense. Although it was a good exercise, it felt a little weird.

In this commit, we return to the more conventional way of creating a new
expense by rendering an `/expenses/new` page. However, we rewrite the form
rendered on that page in React. This is because we want the User select
(labeled with "Paid by") to dynamically update according to which group the
user has selected.

As part of this work, we want to send serialized groups to the front end to
populate the **Group** select element. Currently, the `GroupsSerializer`
serializes the `Group`'s `Expense`s, which is unnecessary here. The current set
up of the `GroupsSerializer` causes us to make an unnecessary query for
`Expense`s.

To solve this, we create an `ApplicationSerializer` with two methods: `#with?`
and `#without?`, which returns a `boolean` for whether a serialized resource
has been requested with or without certain attributes in it's options hash

```ruby
@groups.to_api(without: :expenses).serializable_hash
```

Still to do:
- implement toastr so that we can flash some sort of success/ error message to
  the user when creating a new expense, as the case may be.
- add a validation schema to the formik form