Add ExpenseModal with ExpenseEditForm (#32)
In this commit, we add an `ExpenseModal` component that is triggered by a
button on each `Expense` listed on the `GroupDetails` page.

Clicking this button, opens the modal, where the user sees details of the
expense and another button to edit the given expense.

When clicking the edit button, the contents of the modal get replaced with the
`ExpenseEditForm`, where users can make edits to the expense and submit those
edits or discard them by pressing cancel.

Submitting the form hits the `/api/expenses` endpoint and returns JSON, either
the edited expense or an error message.

Next, we need to implement toaster to show user success/ error messages