Add react-tostify package, display flash as toasts (#24)
In this commit, we add the `react-tostify` package for sending users ephemeral
messages.

We can dispatch toasts from the frontend using the `toast` function.

We also display the contents of the `flash` using `react-tostify`. We mount a
top-level `Toaster` component and pass it the contents of the `flash` as props,
then when the component mounts, we call `flash()` for each of the `toasts`
prop.