Add popover to Expense card on GroupDetails page (#34)
In this commit we update the Expense "card" on the GroupDetails with a button
that triggers a popover. This repalce this button with the `BsBoxArrowUpRight`
icon that opened a modal that showed expense details, then another button to
edit the expense.

In this commit, we replace the `BsBoxArrowUpRight` with a kebab menu that opens
a popover that contains two buttons:
- Edit
- Delete

The Edit button opens the existing ExpenseModal, which has been edited to
render the ExpenseEditForm directly, instead of doing it conditionally.

The next step is to create a modal that is triggered by the Delete button,
which will ask a user to confirm whether they actually want to delete the
expense.