Add DeleteExpenseModal (#34)
In this commit we add a new DeleteExpenseModal which gets triggered when a user
clicks the "Delete" button in the popover for a given expense.

The delete modal displays information about the expense and asks the user to
confirm that they want to delete the expense.

Currently, the **delete** button doesn't do anythign.

In an upcoming commit, we will extend the ExpensesRepository to hit a new
endpoint to delete a given expense.