Add expenses.status column and enum (#34)
In this commit, we add an `expenses.status` column to track the status of an
`Expense`. We also add a status `enum` to the `Expense` model, with three
possible values.

- When users first add an expense, it will be `:open`.
- When users pay the expense, it will change to `:settled`
- When a user deletes the expense, it will change to `:deleted`

The `deleted` status is for the sake of transparency. Instead of completely
destroying an expense after it's been created, we instead change its status and
use that status to exclude it from balances calculations.