Only include open expenses on Group show page, balance calculation (#34)
In preparation for allowing users to delete expenses (and eventually settle
expenses), we want to make sure that only open expenses are included in the
calculation of a group's balance and that we only show open expenses in the
expenses feed on the group show page.

In this commit, we update the `Balances` query to only consider `:open`
expenses when calculating the group's `totalExpense` and each member's balance.

Additionally, in the `GroupDetails` component, we filter for expenses with an
`Open` status and display these in the expenses feed.

Now, we should be set up to allow users to "delete" an expense (i.e., change
its status to `:deleted`) and have this change reflected on the frontend.