Render deleted expenses in accordion (#34)
In this commmit, we render deleted expenses in a "Deleted Expenses" accordion
below the feed of open expenses. The purpose of this is transparency: after an
expense has been added to a group, we want to allow users to delete the expense
and exclude it from the balance calculation and from future settle ups;
however, we don't want a user to delete an expense just because they don't want
to pay it. To prevent this from happening, we don't completely destroy the
expense, we just mark is as deleted.

Additionally, we make changes to the GroupDetailsData component to wrap the
heading and data displayed underneath within a containing `section` element
passed an optional id. This is semantically clearer and allows us to more
easily confirm the data under each heading in tests.