Add Settlements::Create event (#37)
In this commit, we add the `Settlement::Create` event, which will be invoked
when a user settles expenses for the group.

It creates a `Settlement` record and then for all open expenses for the group,
associates the `Settlement` with them and changes the expense's status to
`:settled`.

This is done in a transaction so that if anything goes wrong in the process,
none of the expenses are marked as `:settled`.