Display settlements on group show page (#41)
In this commit, we display settlements in an accordion menu on the group show
page.

In the accordion, we show high level information on each settlement (total,
number of expenses, date of settlement).

When the user clicks the "Show details" button, we render an expenses callout,
a table with the expenses involved in the settlement, and conditionally render
notes regarding the settlement.