Generate Users::RegistrationsController (for Devise) (#44)
Since we set up Devise, we've added new required columns to the `users` table.
This is causing sign up to fail with the out-of-the-box
`Devise::RegistrationsController#create` method.

To solve this, we configure our own controller to handle registrations, per the
Devise docs:

https://github.com/heartcombo/devise?tab=readme-ov-file#configuring-controllers

We'll need to eventually set up some side-effects to a user registering
(generating demo data), which we'll invoke in the #create action, so we take
the approach of creating our own custom controller now rather than permitting
the new users params in `ApplicationController`