Create new group from React new group form (#46)
In this commit, we add the GroupsRepository and an API endpoint to create a new
group.

When the group creation fails, we return JSON errors to the frontend and
display the errors in a toaster.

We also add a `GroupsController#edit` action, as the set up of the GroupForm
component should work with editing an existing group.

Next up: create an update API endpoint to handle edits, update the GroupForm
component to render slightly differently in the editing context.