Allow users to edit group (#48)
In this commit, we add a gear icon next to the group name that directs users to
the /groups/:id/edit page. Once on this page, we render the GroupForm component
with some slight variations to reflect that we are editing an existing group
(add a cancel button, submit button says 'Update Group'). We can tell that
we're editing an existing group by the presence of `group.id` in the group
prop.

We also add an update GroupsRepository method and and update API endpoint for
groups.