Include current_user in Group Members select
In this commit, we include the current_user in the Group Members select on the
Group Form.

When creating the group, we require that the current_user adds themselves to
the group to prevent users from creating groups of which they are not a part.

This also allows the current_user to remove themselves from the group when
editing a group.