Add fields to GroupForm to invite new contacts (#25)
In this commit, we add a new <ArrayField> to the GroupForm component to allow
users to invite new users with whom they are not already in a group to join the
group that is being created/ edited.

This field is optional and the user can generate more than one by clicking a
button. The value in one of the ArrayField inputs is added to an array and will
be sent to the update/ create API endpoints for groups. These will then be used
to create an Invitation, which the user can then accept or decline.