Generate Notification data model (#52)
In this commit, we set up the polymorphic Notification model. Notifcation
belongs to a user (the person who should see the notification) and a source
(the polymorphic association). The source can be an Expense, Settlement, or
Invitation. Notificaton also has a category enum, which describes what the
notification is about.