Notify Invitee of Invitation (#52)
In this commit, we create a Notification for the invitee of a Group Invitation.
We only do this if the invitee exists. It is possible to invite someone who has
not yet signed up for the app. In this case, we will defer the creation of the
notification until the user signs up, then we will associate any orphaned
Notifications with the new user's email with the new user.