Notify group users other than current_user for expense (#57)
Currently, when a user adds or updates an expense, we notfiy all group users,
including the current_user. This creates some noise in the notifications for
the current user. We shouldn't need to notify the user who added or updated an
expense that the expense was added/updated.

In this commit, we change the notification logic to create a notification for
all users except the current user.