Add users.has_seen_demo_modal flag (#69)
In this commit, we add a `has_seen_demo_modal` boolean flag to the `users`
table. We will use this column to determine whether to show the user the About
this demo modal by default.

When false, we show it. When true, we don't show the modal by default, but the
user can view it again by clicking the button in the nav bar.

This value will default to false. When the user dismisses the modal, we'll send
a request to flip this flag to true, indicating that they have viewed the
modal.