Set has_seen_demo_modal to true in Users factory (#69)
The has_seen_demo_modal flag defaults to false for a user, because we want to
show the About this demo modal to new users everytime they log in until they
dismiss the modal, causing its value to be flipped.

However, this was causing a bunch of system spec failures, because the modal
was blocking elements the system spec was expecting to be on the screen.

To get around this, set has_seen_demo_modal to true in the Users factory. In
system specs where we want to test the behavior of the modal/ this flag, we can
either create a new user through the sign up form or set the value of this flag
to false explicitly.