Update SendGrid config
In the previous commit 87ca3031b5186c8b24a079cfd71ce724d485e342 the config was
placed in environment.rb according to SendGrid's docs. This worked in
development, but not in produciton. THis could be because the ENV vars aren't
loaded in time on Render when this config is read in environment.rb.

To attempt to fix this, we move the config to the specific environment files.

This is working in development