Update domain in production smtp_settings
SendGrid still hasn't authenticated the domain billsplit-lylg.onrender.com so i
was using gmail.com as the domain, since I do have billsplit.gf.app@gmail.com
as a verified sender in SendGrid.

However, in production, I'm still getting a Net::OpenTimeout (execution
expired) error when trying to send an email, suggesting that the production app
still can't connect to SendGrid's SMTP server and there is still something
wrong with my SMTP settings.

In this commit, we chage the domain to be the domain of the production app to
see if that makes things better.