Update development smtp config, call deliver_now on InvitationMailer method
We're still not able to send mail from production. I want to test out SendGrid
by ensuring that the same config is used in development and production and
verifying that email does get sent from development.

We weren't calling `deliver_now` on the email to new group invitee, so in this
commit we add that so that the email gets sent right away.

The email was successfully sent from development with the update to the
development config file and the call to `deliver_now`: