Serialize created_at in valid ISO 8601 format
Previously, we were just serializing created_at attributes for Expenses and
Settlements as a stringified version of the ActiveSupport::TimeWithZone class.
This resulted in something like "2025-09-19 14:23:30 UTC"

This worked fine on desktop browsers, but caused the page to crash on mobile
devices when passing such a date to the `format()` function from `date-fns`. To
fix this, we serialize dates in a valid ISO 8601 format.