Add User#destory_demo_data method
In this commit, we add an event to destroy the data created in DemoData:Create.

This is not an optimal solution, but gets the job done in the meantime.

In a future commit, we should update the models to destroy dependent records/
better mark records as being part of a demo.