Add unique index on group_id and user_id to memberships (#51)
This is the first step in solving the issue of a user being able to be a member
of a group twice.

In this commit, we add a unique constraint on the memberships table on user_id
and group_id to prevent a user from having more than one membership in a given
group.

In up coming work, we will add a similar constraint on invitations so that
users won't be able to be invited to the group if they already have an
invitiation.