Add unique index to invitations on email, group, pending (#51)
As part of the work to prevent users from being able to join a group more than
once, we add a database constraint to prevent a user from having more than one
pending invitation for a given group. This will prevent the situation of a user
seeing two invitations for the same group and clicking "Accept" on both of
them.

The unique index is on the invitations table on `group_id` and `invitee_email`,
where the status of the invitation is `pending`. This status component is
important: we want a user to be able to have more than one invitation per group
(say, the user accidentally clicks "decline" and the invitation is marked as
declined; in this case, we would want to be able to send the user another
inviation); however, there should only be one pending invitation at any given
time.