Notify user when trying to create duplicate invite (#51)
In this commit, we notify a user via the flash when they are trying to add a
user to the group who already has a pending invitation.

Since a user may invite multiple users who already have pending invitaitions,
we need to be able to set flash keys with multiple messages. To accomplish
this, we update the _messages partial to handle arrays of messages rather than
only single strings.