Extract formatCurrency util (#72)
To insure that we display expense/ balance amounts in the correct format, we
were using the Intl.NumberFormat.format() function. Calls to this function with
its verbose arguments were repeated throughout the code base.

To reduce the repition and make changes to how we format values (e.g. based on
user preferences) easier in the future, we define a formatCurrency() util and
use this to format currencies.

Additionally, we change the formatting style to be German/ Germany and the
currency to be Euro.