Display archived groups to user (#73)
In this commit, we add a new `/groups/archived` route where a user can go to
view their archived groups.

When displaying archived groups, we use the same `GroupsListContainer` and
`GroupsListCard` components as we do when displaying active groups, but with
slightly different display logic to visually distinguish a group as archived
(an archived pill on the card).

Additionally, we moved the logic in the HomeController#index action to the
GroupsController#index action, as this is where convention dictates this logic
should be. To keep the functionality the same, we update the root_path to route
to `groups#index`